requirements.txt
Main packages used to run the data processing and statistical analysis code (everything except the Simulation folder):

cellpose==3.0.11
scikit-image==0.24.0
scipy==1.14.1
pyspm==0.6.2
pillow==10.4.0
matplotlib==3.9.2
tqdm==4.66.5
numba==0.60.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
scikit-learn==1.5.2
natsort==8.4.0

Main packages used to run the simulation code (use conda env):
fenics-dolfinx=0.9
pyvista=0.44
mpich=4.2.3
numpy=2.1.3
gmsh=4.13.1
matplotlib=3.9.3
scipy=1.14.1
petsc=3.22.2
petsc4py=3.22.1
mpi4py=4.0.1
